Completed SimulatorPanel (-integration and Stage)
Finished layout for the Simulator panel, not including the Stage where
the maze is drawn-- currently only a JTextArea to fill the space.
This commit is mostly to give me a restore point before I start swinging
at integrating this into the main program, and putting the Stage +
Simulator components into the SimPanel.